+++ pza/docs/CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
 ## Version 0.4.2
+
 - Updated README.md documentation
 - Changed license to MIT
 - Added placeholder project logo image
+
+## Version 0.4.6
+
+- Updated version number
+- Changed logo
+- Updated folder path helpers to return folder paths ending with separators
+- Added DatabaseHub.connection attribute
+- Added SettingsHub.profile_settings_file_name attribute
+- Added I/O helpers to read/write list files
+- Standardize double-quotes for string literals
+- Added data type validation module
+- Added database helpers to generate sql query from a dictionary
+- Fixed pride project.scripts in project.toml
+- Updated README
+- Added code files for README examples